test: split dashboard server and context tests
