+++ tests/test_store_dashboard_mcp.py
@@ -2,13 +2,6 @@ from __future__ import annotations
 
 import json
 import sqlite3
-import sys
-import threading
-import urllib.error
-import urllib.parse
-import urllib.request
-from functools import partial
-from http.server import ThreadingHTTPServer
 from pathlib import Path
 from typing import Any
 
@@ -16,13 +9,9 @@ import pytest
 from store_dashboard_helpers import (
     SECOND_SESSION_ID,
     SESSION_ID,
-    _assert_contract,
     _entry,
     _extract_js_function,
-    _fake_pricing_update,
-    _http_error_json,
     _make_codex_home,
-    _read_json,
     _token_event,
     _usage_event,
     _write_archived_log,
@@ -31,14 +20,8 @@ from store_dashboard_helpers import (
 )
 
 from codex_usage_tracker import store as store_module
-from codex_usage_tracker.context import load_call_context
 from codex_usage_tracker.dashboard import dashboard_payload, generate_dashboard
-from codex_usage_tracker.diagnostics import run_doctor
 from codex_usage_tracker.models import UsageEvent
-from codex_usage_tracker.pricing import (
-    annotate_rows_with_efficiency,
-    load_pricing_config,
-)
 from codex_usage_tracker.store import (
     EVENT_COLUMNS,
     connect,
@@ -1154,844 +1137,6 @@ def test_dashboard_guide_link_can_use_docs_url_override(tmp_path: Path, monkeypa
     assert not (tmp_path / "codex-usage-tracker-guide").exists()
     assert (tmp_path / "codex-usage-tracker-assets" / "dashboard.js").exists()
 
-
-def test_dashboard_server_usage_api_refreshes_aggregate_rows(tmp_path: Path) -> None:
-    from codex_usage_tracker.server import _UsageDashboardHandler
-
-    codex_home = _make_codex_home(tmp_path)
-    db_path = tmp_path / "usage.sqlite3"
-    pricing_path = _write_pricing(tmp_path / "pricing.json")
-    (tmp_path / "dashboard.html").write_text("<!doctype html><title>Dashboard</title>", encoding="utf-8")
-    handler = partial(
-        _UsageDashboardHandler,
-        directory=str(tmp_path),
-        db_path=db_path,
-        pricing_path=pricing_path,
-        allowance_path=tmp_path / "allowance.json",
-        thresholds_path=tmp_path / "thresholds.json",
-        projects_path=tmp_path / "projects.json",
-        limit=5000,
-        since=None,
-        codex_home=codex_home,
-        include_archived=False,
-        dashboard_name="dashboard.html",
-        context_chars=2000,
-        api_token="test-token",
-        context_api_enabled=True,
-        refresh_lock=threading.Lock(),
-    )
-    server = ThreadingHTTPServer(("127.0.0.1", 0), handler)
-    thread = threading.Thread(target=server.serve_forever, daemon=True)
-    thread.start()
-    try:
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            f"http://127.0.0.1:{server.server_port}/dashboard.html",
-            timeout=5,
-        ) as response:
-            dashboard_cache_control = response.headers.get("Cache-Control")
-            dashboard_html = response.read().decode("utf-8")
-        refresh_without_token = _http_error_json(
-            f"http://127.0.0.1:{server.server_port}/api/usage?refresh=1&limit=2"
-        )
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            urllib.request.Request(
-                f"http://127.0.0.1:{server.server_port}/api/usage?refresh=1&limit=2",
-                headers={"X-Codex-Usage-Token": "test-token"},
-            ),
-            timeout=5,
-        ) as response:
-            content_security_policy = response.headers.get("Content-Security-Policy")
-            referrer_policy = response.headers.get("Referrer-Policy")
-            limited_payload = json.loads(response.read().decode("utf-8"))
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            f"http://127.0.0.1:{server.server_port}/api/usage?limit=all",
-            timeout=5,
-        ) as response:
-            all_payload = json.loads(response.read().decode("utf-8"))
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            f"http://127.0.0.1:{server.server_port}/api/usage?limit=2&offset=2",
-            timeout=5,
-        ) as response:
-            offset_payload = json.loads(response.read().decode("utf-8"))
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            f"http://127.0.0.1:{server.server_port}/api/usage?shell=1&limit=all",
-            timeout=5,
-        ) as response:
-            shell_payload = json.loads(response.read().decode("utf-8"))
-        forbidden_origin = _http_error_json(
-            f"http://127.0.0.1:{server.server_port}/api/usage",
-            headers={"Origin": "http://example.test"},
-        )
-    finally:
-        server.shutdown()
-        server.server_close()
-        thread.join(timeout=5)
-
-    assert refresh_without_token["status"] == 403
-    assert dashboard_cache_control == "no-store"
-    shell_raw_payload = dashboard_html.split(
-        '<script id="usage-data" type="application/json">',
-        1,
-    )[1].split("</script>", 1)[0]
-    dashboard_shell_payload = json.loads(shell_raw_payload)
-    assert 'data-dashboard-shell="true"' in dashboard_html
-    assert dashboard_shell_payload["shell_boot"] is True
-    assert dashboard_shell_payload["rows"] == []
-    assert dashboard_shell_payload["summary"]["visible_calls"] == 0
-    assert limited_payload["refresh_result"]["parsed_events"] == 4
-    assert limited_payload["refresh_result"]["skipped_events"] == 0
-    assert limited_payload["refresh_result"]["parser_diagnostics"] == {}
-    assert len(limited_payload["rows"]) == 2
-    assert limited_payload["loaded_row_count"] == 2
-    assert limited_payload["total_available_rows"] == 4
-    assert limited_payload["limit"] == 2
-    assert limited_payload["offset"] == 0
-    assert limited_payload["has_more"] is True
-    assert limited_payload["next_offset"] == 2
-    assert content_security_policy is not None
-    assert "connect-src 'self'" in content_security_policy
-    assert "unsafe-inline" not in content_security_policy
-    assert referrer_policy == "no-referrer"
-    assert len(all_payload["rows"]) == 4
-    assert all_payload["loaded_row_count"] == 4
-    assert all_payload["total_available_rows"] == 4
-    assert all_payload["limit"] is None
-    assert all_payload["offset"] == 0
-    assert all_payload["has_more"] is False
-    assert all_payload["limit_label"] == "All"
-    assert len(offset_payload["rows"]) == 2
-    assert shell_payload["shell_boot"] is True
-    assert shell_payload["rows"] == []
-    assert shell_payload["summary"]["visible_calls"] == 4
-    assert shell_payload["summary"]["total_tokens"] == sum(
-        row["total_tokens"] for row in all_payload["rows"]
-    )
-    assert shell_payload["limit"] is None
-    assert shell_payload["loaded_row_count"] == 0
-    assert offset_payload["loaded_row_count"] == 2
-    assert offset_payload["total_available_rows"] == 4
-    assert offset_payload["limit"] == 2
-    assert offset_payload["offset"] == 2
-    assert offset_payload["has_more"] is False
-    assert offset_payload["next_offset"] is None
-    assert {row["record_id"] for row in offset_payload["rows"]}.isdisjoint(
-        {row["record_id"] for row in limited_payload["rows"]}
-    )
-    assert limited_payload["pricing_configured"] is True
-    assert limited_payload["allowance_configured"] is False
-    assert limited_payload["allowance_source"]["name"] == "OpenAI Codex rate card"
-    assert limited_payload["rows"][0]["usage_credits"] is not None
-    assert "refreshed_at" in limited_payload
-    assert limited_payload["parser_diagnostics"] == {}
-    assert limited_payload["api_token"] == "test-token"
-    assert limited_payload["context_api_enabled"] is True
-    assert forbidden_origin["status"] == 403
-    assert "SECRET RAW PROMPT" not in json.dumps(limited_payload)
-
-
-def test_dashboard_history_scope_excludes_archived_rows_by_default(tmp_path: Path) -> None:
-    codex_home = _make_codex_home(tmp_path)
-    _write_archived_log(codex_home)
-    db_path = tmp_path / "usage.sqlite3"
-    refresh_result = refresh_usage_index(
-        codex_home=codex_home,
-        db_path=db_path,
-        include_archived=True,
-    )
-
-    active_payload = dashboard_payload(db_path=db_path, limit=0)
-    all_history_payload = dashboard_payload(db_path=db_path, limit=0, include_archived=True)
-    active_rows = query_dashboard_events(db_path=db_path, limit=0, include_archived=False)
-    all_rows = query_dashboard_events(db_path=db_path, limit=0, include_archived=True)
-    archived_rows = [row for row in all_rows if "/archived_sessions/" in row["source_file"]]
-
-    assert refresh_result.parsed_events == 5
-    assert active_payload["include_archived"] is False
-    assert active_payload["history_scope"] == "active"
-    assert active_payload["loaded_row_count"] == 4
-    assert active_payload["total_available_rows"] == 4
-    assert active_payload["active_available_rows"] == 4
-    assert active_payload["all_history_available_rows"] == 5
-    assert active_payload["archived_available_rows"] == 1
-    assert all_history_payload["include_archived"] is True
-    assert all_history_payload["history_scope"] == "all-history"
-    assert all_history_payload["loaded_row_count"] == 5
-    assert all_history_payload["total_available_rows"] == 5
-    assert len(active_rows) == 4
-    assert len(all_rows) == 5
-    assert not any("/archived_sessions/" in row["source_file"] for row in active_rows)
-    assert len(archived_rows) == 1
-    assert archived_rows[0]["is_archived"] == 1
-    assert all(row["is_archived"] == 0 for row in active_rows)
-
-    with connect(db_path) as conn:
-        conn.execute("UPDATE usage_events SET is_archived = 0")
-    active_rows_after_migrated_flag_reset = query_dashboard_events(
-        db_path=db_path,
-        limit=0,
-        include_archived=False,
-    )
-    assert len(active_rows_after_migrated_flag_reset) == 4
-    assert not any(
-        "/archived_sessions/" in row["source_file"]
-        for row in active_rows_after_migrated_flag_reset
-    )
-
-
-def test_dashboard_server_usage_api_switches_history_scope(tmp_path: Path) -> None:
-    from codex_usage_tracker.server import _UsageDashboardHandler
-
-    codex_home = _make_codex_home(tmp_path)
-    _write_archived_log(codex_home)
-    db_path = tmp_path / "usage.sqlite3"
-    pricing_path = _write_pricing(tmp_path / "pricing.json")
-    handler = partial(
-        _UsageDashboardHandler,
-        directory=str(tmp_path),
-        db_path=db_path,
-        pricing_path=pricing_path,
-        allowance_path=tmp_path / "allowance.json",
-        thresholds_path=tmp_path / "thresholds.json",
-        projects_path=tmp_path / "projects.json",
-        limit=5000,
-        since=None,
-        codex_home=codex_home,
-        include_archived=False,
-        dashboard_name="dashboard.html",
-        context_chars=2000,
-        api_token="test-token",
-        context_api_enabled=True,
-        refresh_lock=threading.Lock(),
-    )
-    server = ThreadingHTTPServer(("127.0.0.1", 0), handler)
-    thread = threading.Thread(target=server.serve_forever, daemon=True)
-    thread.start()
-    try:
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            urllib.request.Request(
-                f"http://127.0.0.1:{server.server_port}/api/usage?refresh=1&limit=all",
-                headers={"X-Codex-Usage-Token": "test-token"},
-            ),
-            timeout=5,
-        ) as response:
-            active_payload = json.loads(response.read().decode("utf-8"))
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            urllib.request.Request(
-                f"http://127.0.0.1:{server.server_port}/api/usage?refresh=1&limit=all&include_archived=1",
-                headers={"X-Codex-Usage-Token": "test-token"},
-            ),
-            timeout=5,
-        ) as response:
-            all_history_payload = json.loads(response.read().decode("utf-8"))
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            f"http://127.0.0.1:{server.server_port}/api/usage?limit=all&include_archived=0",
-            timeout=5,
-        ) as response:
-            active_after_all_payload = json.loads(response.read().decode("utf-8"))
-    finally:
-        server.shutdown()
-        server.server_close()
-        thread.join(timeout=5)
-
-    assert active_payload["include_archived"] is False
-    assert active_payload["loaded_row_count"] == 4
-    assert active_payload["archived_available_rows"] == 0
-    assert active_payload["refresh_result"]["include_archived"] is False
-    assert all_history_payload["include_archived"] is True
-    assert all_history_payload["loaded_row_count"] == 5
-    assert all_history_payload["archived_available_rows"] == 1
-    assert all_history_payload["refresh_result"]["include_archived"] is True
-    assert active_after_all_payload["include_archived"] is False
-    assert active_after_all_payload["loaded_row_count"] == 4
-    assert active_after_all_payload["archived_available_rows"] == 1
-    assert not any(
-        "/archived_sessions/" in row["source_file"]
-        for row in active_after_all_payload["rows"]
-    )
-
-
-def test_dashboard_server_api_timing_diagnostics_are_opt_in_and_technical(
-    tmp_path: Path,
-) -> None:
-    from codex_usage_tracker.server import _UsageDashboardHandler
-
-    codex_home = _make_codex_home(tmp_path)
-    db_path = tmp_path / "usage.sqlite3"
-    pricing_path = _write_pricing(tmp_path / "pricing.json")
-    handler = partial(
-        _UsageDashboardHandler,
-        directory=str(tmp_path),
-        db_path=db_path,
-        pricing_path=pricing_path,
-        allowance_path=tmp_path / "allowance.json",
-        thresholds_path=tmp_path / "thresholds.json",
-        projects_path=tmp_path / "projects.json",
-        limit=5000,
-        since=None,
-        codex_home=codex_home,
-        include_archived=False,
-        dashboard_name="dashboard.html",
-        context_chars=2000,
-        api_token="test-token",
-        context_api_enabled=True,
-        refresh_lock=threading.Lock(),
-    )
-    server = ThreadingHTTPServer(("127.0.0.1", 0), handler)
-    thread = threading.Thread(target=server.serve_forever, daemon=True)
-    thread.start()
-    try:
-        base_url = f"http://127.0.0.1:{server.server_port}"
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            urllib.request.Request(
-                f"{base_url}/api/usage?refresh=1&limit=2&diagnostics=true",
-                headers={"X-Codex-Usage-Token": "test-token"},
-            ),
-            timeout=5,
-        ) as response:
-            usage_diagnostics_payload = json.loads(response.read().decode("utf-8"))
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            f"{base_url}/api/usage?limit=2",
-            timeout=5,
-        ) as response:
-            usage_default_payload = json.loads(response.read().decode("utf-8"))
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            f"{base_url}/api/usage?limit=2&diagnostics=false",
-            timeout=5,
-        ) as response:
-            usage_false_payload = json.loads(response.read().decode("utf-8"))
-
-        record_id = usage_diagnostics_payload["rows"][0]["record_id"]
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            urllib.request.Request(
-                f"{base_url}/api/context?record_id={record_id}&diagnostics=true",
-                headers={"X-Codex-Usage-Token": "test-token"},
-            ),
-            timeout=5,
-        ) as response:
-            context_diagnostics_payload = json.loads(response.read().decode("utf-8"))
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            urllib.request.Request(
-                f"{base_url}/api/context?record_id={record_id}",
-                headers={"X-Codex-Usage-Token": "test-token"},
-            ),
-            timeout=5,
-        ) as response:
-            context_default_payload = json.loads(response.read().decode("utf-8"))
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            urllib.request.Request(
-                f"{base_url}/api/context?record_id={record_id}&diagnostics=false",
-                headers={"X-Codex-Usage-Token": "test-token"},
-            ),
-            timeout=5,
-        ) as response:
-            context_false_payload = json.loads(response.read().decode("utf-8"))
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            urllib.request.Request(
-                f"{base_url}/api/context?record_id={record_id}&mode=full",
-                headers={"X-Codex-Usage-Token": "test-token"},
-            ),
-            timeout=5,
-        ) as response:
-            context_full_payload = json.loads(response.read().decode("utf-8"))
-        invalid_context_mode = _http_error_json(
-            f"{base_url}/api/context?record_id={record_id}&mode=slow",
-            headers={"X-Codex-Usage-Token": "test-token"},
-        )
-    finally:
-        server.shutdown()
-        server.server_close()
-        thread.join(timeout=5)
-
-    assert "diagnostics" not in usage_default_payload
-    assert "diagnostics" not in usage_false_payload
-    usage_diagnostics = usage_diagnostics_payload["diagnostics"]
-    assert usage_diagnostics["refresh_ms"] >= 0
-    assert usage_diagnostics["dashboard_payload_ms"] >= 0
-    assert usage_diagnostics["json_bytes"] > 0
-    assert usage_diagnostics["rows_returned"] == 2
-    assert usage_diagnostics["include_archived"] is False
-    assert usage_diagnostics["limit"] == 2
-    assert usage_diagnostics["offset"] == 0
-    usage_diagnostics_text = json.dumps(usage_diagnostics)
-    assert "SECRET RAW PROMPT" not in usage_diagnostics_text
-    assert "source_file" not in usage_diagnostics_text
-
-    assert "diagnostics" not in context_default_payload
-    assert "diagnostics" not in context_false_payload
-    assert context_default_payload["context_mode"] == "quick"
-    assert context_default_payload["serialized_evidence"]["deferred_buckets"] is True
-    assert context_full_payload["context_mode"] == "full"
-    assert context_full_payload["serialized_evidence"]["deferred_buckets"] is False
-    assert invalid_context_mode["status"] == 400
-    context_diagnostics = context_diagnostics_payload["diagnostics"]
-    assert context_diagnostics["db_lookup_ms"] >= 0
-    assert context_diagnostics["source_scan_ms"] >= 0
-    assert context_diagnostics["serialized_estimate_ms"] >= 0
-    assert context_diagnostics["source_file_bytes"] > 0
-    assert context_diagnostics["source_line_number"] > 0
-    assert context_diagnostics["entries_before_limit"] >= context_diagnostics["entries_returned"]
-    assert context_diagnostics["json_bytes"] > 0
-    context_diagnostics_text = json.dumps(context_diagnostics)
-    assert "SECRET RAW PROMPT" not in context_diagnostics_text
-    assert str(codex_home) not in context_diagnostics_text
-    assert ".jsonl" not in context_diagnostics_text
-
-
-def test_dashboard_server_live_sql_api_slices_are_aggregate_only(tmp_path: Path) -> None:
-    from codex_usage_tracker.server import _UsageDashboardHandler
-
-    codex_home = _make_codex_home(tmp_path)
-    db_path = tmp_path / "usage.sqlite3"
-    pricing_path = _write_pricing(tmp_path / "pricing.json")
-    refresh_usage_index(codex_home=codex_home, db_path=db_path)
-    handler = partial(
-        _UsageDashboardHandler,
-        directory=str(tmp_path),
-        db_path=db_path,
-        pricing_path=pricing_path,
-        allowance_path=tmp_path / "allowance.json",
-        thresholds_path=tmp_path / "thresholds.json",
-        projects_path=tmp_path / "projects.json",
-        limit=5000,
-        since=None,
-        codex_home=codex_home,
-        include_archived=False,
-        dashboard_name="dashboard.html",
-        context_chars=2000,
-        api_token="test-token",
-        context_api_enabled=True,
-        refresh_lock=threading.Lock(),
-    )
-    server = ThreadingHTTPServer(("127.0.0.1", 0), handler)
-    thread = threading.Thread(target=server.serve_forever, daemon=True)
-    thread.start()
-    try:
-        base_url = f"http://127.0.0.1:{server.server_port}"
-
-        status_payload = _read_json(f"{base_url}/api/status")
-        calls_payload = _read_json(
-            f"{base_url}/api/calls?limit=2&sort=tokens&direction=desc&q=Codex"
-        )
-        offset_payload = _read_json(
-            f"{base_url}/api/calls?limit=1&offset=1&sort=tokens&direction=desc&q=Codex"
-        )
-        record_id = calls_payload["rows"][0]["record_id"]
-        call_payload = _read_json(f"{base_url}/api/call?record_id={record_id}")
-        threads_payload = _read_json(f"{base_url}/api/threads?limit=2&sort=tokens")
-        thread_key = threads_payload["rows"][0]["thread_key"]
-        thread_calls_payload = _read_json(
-            f"{base_url}/api/thread-calls?thread_key={urllib.parse.quote(thread_key)}&limit=2"
-        )
-        summary_payload = _read_json(f"{base_url}/api/summary?group_by=model&limit=5")
-        recommendations_payload = _read_json(f"{base_url}/api/recommendations?limit=5")
-        invalid_sort = _http_error_json(f"{base_url}/api/calls?sort=not-a-sort")
-    finally:
-        server.shutdown()
-        server.server_close()
-        thread.join(timeout=5)
-
-    assert status_payload["schema"] == "codex-usage-tracker-status-v1"
-    _assert_contract(status_payload)
-    assert "rows" not in status_payload
-    assert status_payload["row_counts"]["scoped_rows"] == 4
-    assert status_payload["max_event_timestamp"]
-
-    assert calls_payload["schema"] == "codex-usage-tracker-calls-v1"
-    _assert_contract(calls_payload)
-    assert calls_payload["row_count"] == 2
-    assert calls_payload["total_matched_rows"] >= 2
-    assert calls_payload["has_more"] is True
-    assert calls_payload["next_offset"] == 2
-    assert calls_payload["rows"][0]["total_tokens"] >= calls_payload["rows"][1]["total_tokens"]
-    assert offset_payload["rows"][0]["record_id"] != calls_payload["rows"][0]["record_id"]
-
-    assert call_payload["schema"] == "codex-usage-tracker-call-v1"
-    _assert_contract(call_payload)
-    assert call_payload["record"]["record_id"] == record_id
-    assert call_payload["raw_context_included"] is False
-
-    assert threads_payload["schema"] == "codex-usage-tracker-threads-v1"
-    _assert_contract(threads_payload)
-    assert threads_payload["row_count"] >= 1
-    assert "total_tokens" in threads_payload["rows"][0]
-
-    assert thread_calls_payload["schema"] == "codex-usage-tracker-thread-calls-v1"
-    _assert_contract(thread_calls_payload)
-    assert thread_calls_payload["thread_key"] == thread_key
-    assert thread_calls_payload["row_count"] >= 1
-
-    assert summary_payload["schema"] == "codex-usage-tracker-summary-v1"
-    _assert_contract(summary_payload)
-    assert summary_payload["group_by"] == "model"
-    assert recommendations_payload["schema"] == "codex-usage-tracker-recommendations-v1"
-    _assert_contract(recommendations_payload)
-    assert invalid_sort["status"] == 400
-    assert "sort must be one of" in invalid_sort["payload"]["error"]
-
-    combined_payload = json.dumps(
-        [
-            status_payload,
-            calls_payload,
-            call_payload,
-            threads_payload,
-            thread_calls_payload,
-            summary_payload,
-            recommendations_payload,
-        ]
-    )
-    assert "SECRET RAW PROMPT" not in combined_payload
-    assert "raw_context_included\": true" not in combined_payload
-
-
-def test_dashboard_server_serves_lightweight_call_investigator_boot_html(
-    tmp_path: Path,
-) -> None:
-    from codex_usage_tracker.server import _UsageDashboardHandler
-
-    codex_home = _make_codex_home(tmp_path)
-    db_path = tmp_path / "usage.sqlite3"
-    pricing_path = _write_pricing(tmp_path / "pricing.json")
-    refresh_usage_index(codex_home=codex_home, db_path=db_path)
-    dashboard_path = tmp_path / "dashboard.html"
-    generate_dashboard(
-        db_path=db_path,
-        output_path=dashboard_path,
-        pricing_path=pricing_path,
-        api_token="test-token",
-        context_api_enabled=True,
-        include_archived=True,
-    )
-    record_id = query_dashboard_events(db_path=db_path, include_archived=True)[0]["record_id"]
-    static_html = dashboard_path.read_text(encoding="utf-8")
-    handler = partial(
-        _UsageDashboardHandler,
-        directory=str(tmp_path),
-        db_path=db_path,
-        pricing_path=pricing_path,
-        allowance_path=tmp_path / "allowance.json",
-        thresholds_path=tmp_path / "thresholds.json",
-        projects_path=tmp_path / "projects.json",
-        limit=5000,
-        since=None,
-        codex_home=codex_home,
-        include_archived=False,
-        dashboard_name="dashboard.html",
-        dashboard_path=dashboard_path,
-        context_chars=2000,
-        api_token="test-token",
-        context_api_enabled=True,
-        refresh_lock=threading.Lock(),
-    )
-    server = ThreadingHTTPServer(("127.0.0.1", 0), handler)
-    thread = threading.Thread(target=server.serve_forever, daemon=True)
-    thread.start()
-    try:
-        url = (
-            f"http://127.0.0.1:{server.server_port}/dashboard.html"
-            f"?view=call&record={urllib.parse.quote(record_id)}&history=all"
-        )
-        with urllib.request.urlopen(url, timeout=5) as response:  # noqa: S310 - local test server only
-            html = response.read().decode("utf-8")
-    finally:
-        server.shutdown()
-        server.server_close()
-        thread.join(timeout=5)
-
-    raw_payload = html.split('<script id="usage-data" type="application/json">', 1)[1].split(
-        "</script>",
-        1,
-    )[0]
-    payload = json.loads(raw_payload)
-    assert 'data-active-view="call"' in html
-    assert 'data-investigator-boot="true"' in html
-    assert 'data-dashboard-shell="true"' in html
-    assert payload["investigator_boot"] is True
-    assert payload["rows"] == []
-    assert payload["include_archived"] is True
-    assert payload["api_token"] == "test-token"
-    assert len(html) < len(static_html)
-
-
-def test_dashboard_call_api_returns_adjacent_records_for_investigator(
-    tmp_path: Path,
-) -> None:
-    from codex_usage_tracker.server import _UsageDashboardHandler
-
-    db_path = tmp_path / "usage.sqlite3"
-    events = [
-        _usage_event(
-            record_id="a1",
-            session_id="session-a",
-            thread_key="thread:Alpha",
-            event_timestamp="2026-05-17T18:58:20Z",
-            cumulative_total_tokens=100,
-        ),
-        _usage_event(
-            record_id="a2",
-            session_id="session-a",
-            thread_key="thread:Alpha",
-            event_timestamp="2026-05-17T18:58:30Z",
-            cumulative_total_tokens=250,
-        ),
-        _usage_event(
-            record_id="a3",
-            session_id="session-a",
-            thread_key="thread:Alpha",
-            event_timestamp="2026-05-17T18:58:40Z",
-            cumulative_total_tokens=400,
-        ),
-    ]
-    upsert_usage_events(events, db_path=db_path)
-    handler = partial(
-        _UsageDashboardHandler,
-        directory=str(tmp_path),
-        db_path=db_path,
-        pricing_path=_write_pricing(tmp_path / "pricing.json"),
-        allowance_path=tmp_path / "allowance.json",
-        thresholds_path=tmp_path / "thresholds.json",
-        projects_path=tmp_path / "projects.json",
-        limit=5000,
-        since=None,
-        codex_home=tmp_path / ".codex",
-        include_archived=False,
-        dashboard_name="dashboard.html",
-        context_chars=2000,
-        api_token="test-token",
-        context_api_enabled=True,
-        refresh_lock=threading.Lock(),
-    )
-    server = ThreadingHTTPServer(("127.0.0.1", 0), handler)
-    thread = threading.Thread(target=server.serve_forever, daemon=True)
-    thread.start()
-    try:
-        payload = _read_json(
-            f"http://127.0.0.1:{server.server_port}/api/call?record_id=a2",
-            headers={"X-Codex-Usage-Token": "test-token"},
-        )
-    finally:
-        server.shutdown()
-        server.server_close()
-        thread.join(timeout=5)
-
-    assert payload["schema"] == "codex-usage-tracker-call-v1"
-    assert payload["record"]["record_id"] == "a2"
-    assert payload["previous_record"]["record_id"] == "a1"
-    assert payload["next_record"]["record_id"] == "a3"
-    assert [row["record_id"] for row in payload["adjacent_records"]] == ["a1", "a2", "a3"]
-    assert payload["raw_context_included"] is False
-
-
-def test_dashboard_server_opens_only_token_protected_investigator_urls(
-    tmp_path: Path, monkeypatch
-) -> None:
-    from codex_usage_tracker import server as server_module
-    from codex_usage_tracker.server import _UsageDashboardHandler
-
-    opened_urls: list[str] = []
-    monkeypatch.setattr(server_module.webbrowser, "open_new_tab", lambda url: opened_urls.append(url) or True)
-    (tmp_path / "dashboard.html").write_text("<!doctype html><title>Dashboard</title>", encoding="utf-8")
-    handler = partial(
-        _UsageDashboardHandler,
-        directory=str(tmp_path),
-        db_path=tmp_path / "usage.sqlite3",
-        pricing_path=tmp_path / "pricing.json",
-        allowance_path=tmp_path / "allowance.json",
-        thresholds_path=tmp_path / "thresholds.json",
-        projects_path=tmp_path / "projects.json",
-        limit=5000,
-        since=None,
-        codex_home=_make_codex_home(tmp_path),
-        include_archived=False,
-        dashboard_name="dashboard.html",
-        context_chars=2000,
-        api_token="test-token",
-        context_api_enabled=True,
-        refresh_lock=threading.Lock(),
-    )
-    server = ThreadingHTTPServer(("127.0.0.1", 0), handler)
-    thread = threading.Thread(target=server.serve_forever, daemon=True)
-    thread.start()
-    try:
-        base_url = f"http://127.0.0.1:{server.server_port}"
-        target = f"{base_url}/dashboard.html?view=call&record=abc123&history=all"
-        encoded_target = urllib.parse.quote(target, safe="")
-        without_token = _http_error_json(
-            f"{base_url}/api/open-investigator?url={encoded_target}"
-        )
-        external_url = urllib.parse.quote("https://example.test/dashboard.html?view=call&record=abc123", safe="")
-        external_error = _http_error_json(
-            f"{base_url}/api/open-investigator?url={external_url}",
-            headers={"X-Codex-Usage-Token": "test-token"},
-        )
-        missing_record = urllib.parse.quote(f"{base_url}/dashboard.html?view=call", safe="")
-        missing_record_error = _http_error_json(
-            f"{base_url}/api/open-investigator?url={missing_record}",
-            headers={"X-Codex-Usage-Token": "test-token"},
-        )
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            urllib.request.Request(
-                f"{base_url}/api/open-investigator?url={encoded_target}",
-                headers={"X-Codex-Usage-Token": "test-token"},
-            ),
-            timeout=5,
-        ) as response:
-            payload = json.loads(response.read().decode("utf-8"))
-    finally:
-        server.shutdown()
-        server.server_close()
-        thread.join(timeout=5)
-
-    assert without_token["status"] == 403
-    assert external_error["status"] == 400
-    assert missing_record_error["status"] == 400
-    assert payload["schema"] == "codex-usage-tracker-open-investigator-v1"
-    assert payload["opened"] is True
-    assert opened_urls == [target]
-
-
-def test_dashboard_server_returns_json_for_sqlite_errors(tmp_path: Path, monkeypatch) -> None:
-    from codex_usage_tracker import server as server_module
-    from codex_usage_tracker.server import _UsageDashboardHandler
-
-    def broken_dashboard_payload(**kwargs):
-        raise sqlite3.OperationalError("database is locked")
-
-    def broken_context(**kwargs):
-        raise sqlite3.OperationalError("database is locked")
-
-    monkeypatch.setattr(server_module, "dashboard_payload", broken_dashboard_payload)
-    monkeypatch.setattr(server_module, "load_call_context", broken_context)
-    handler = partial(
-        _UsageDashboardHandler,
-        directory=str(tmp_path),
-        db_path=tmp_path / "usage.sqlite3",
-        pricing_path=tmp_path / "pricing.json",
-        allowance_path=tmp_path / "allowance.json",
-        thresholds_path=tmp_path / "thresholds.json",
-        projects_path=tmp_path / "projects.json",
-        limit=5000,
-        since=None,
-        codex_home=tmp_path / ".codex",
-        include_archived=False,
-        dashboard_name="dashboard.html",
-        context_chars=2000,
-        api_token="test-token",
-        context_api_enabled=True,
-        refresh_lock=threading.Lock(),
-    )
-    server = ThreadingHTTPServer(("127.0.0.1", 0), handler)
-    thread = threading.Thread(target=server.serve_forever, daemon=True)
-    thread.start()
-    try:
-        usage_error = _http_error_json(
-            f"http://127.0.0.1:{server.server_port}/api/usage"
-        )
-        context_error = _http_error_json(
-            f"http://127.0.0.1:{server.server_port}/api/context?record_id=abc",
-            headers={"X-Codex-Usage-Token": "test-token"},
-        )
-    finally:
-        server.shutdown()
-        server.server_close()
-        thread.join(timeout=5)
-
-    assert usage_error["status"] == 500
-    assert "Database error" in usage_error["payload"]["error"]
-    assert context_error["status"] == 500
-    assert "Database error" in context_error["payload"]["error"]
-
-
-def test_dashboard_server_can_enable_context_api_at_runtime(tmp_path: Path) -> None:
-    from codex_usage_tracker.server import _ContextApiState, _UsageDashboardHandler
-
-    codex_home = _make_codex_home(tmp_path)
-    db_path = tmp_path / "usage.sqlite3"
-    refresh_usage_index(codex_home=codex_home, db_path=db_path)
-    record_id = query_session_usage(db_path=db_path, session_id=SESSION_ID)[0]["record_id"]
-    context_api_state = _ContextApiState(False)
-
-    handler = partial(
-        _UsageDashboardHandler,
-        directory=str(tmp_path),
-        db_path=db_path,
-        pricing_path=tmp_path / "pricing.json",
-        allowance_path=tmp_path / "allowance.json",
-        thresholds_path=tmp_path / "thresholds.json",
-        projects_path=tmp_path / "projects.json",
-        limit=5000,
-        since=None,
-        codex_home=codex_home,
-        include_archived=False,
-        dashboard_name="dashboard.html",
-        context_chars=2000,
-        api_token="test-token",
-        context_api_state=context_api_state,
-        refresh_lock=threading.Lock(),
-    )
-    server = ThreadingHTTPServer(("127.0.0.1", 0), handler)
-    thread = threading.Thread(target=server.serve_forever, daemon=True)
-    thread.start()
-    try:
-        disabled_error = _http_error_json(
-            f"http://127.0.0.1:{server.server_port}/api/context?record_id={record_id}",
-            headers={"X-Codex-Usage-Token": "test-token"},
-        )
-        enable_without_token = _http_error_json(
-            f"http://127.0.0.1:{server.server_port}/api/context-settings?enabled=1"
-        )
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            urllib.request.Request(
-                f"http://127.0.0.1:{server.server_port}/api/context-settings?enabled=1",
-                headers={"X-Codex-Usage-Token": "test-token"},
-            ),
-            timeout=5,
-        ) as response:
-            settings_payload = json.loads(response.read().decode("utf-8"))
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            urllib.request.Request(
-                f"http://127.0.0.1:{server.server_port}/api/usage?limit=1",
-                headers={"X-Codex-Usage-Token": "test-token"},
-            ),
-            timeout=5,
-        ) as response:
-            usage_payload = json.loads(response.read().decode("utf-8"))
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            urllib.request.Request(
-                f"http://127.0.0.1:{server.server_port}/api/context?record_id={record_id}",
-                headers={"X-Codex-Usage-Token": "test-token"},
-            ),
-            timeout=5,
-        ) as response:
-            context_payload = json.loads(response.read().decode("utf-8"))
-        with urllib.request.urlopen(  # noqa: S310 - local test server only
-            urllib.request.Request(
-                f"http://127.0.0.1:{server.server_port}/api/context?record_id={record_id}&max_chars=0&max_entries=0",
-                headers={"X-Codex-Usage-Token": "test-token"},
-            ),
-            timeout=5,
-        ) as response:
-            unlimited_context_payload = json.loads(response.read().decode("utf-8"))
-    finally:
-        server.shutdown()
-        server.server_close()
-        thread.join(timeout=5)
-
-    assert disabled_error["status"] == 403
-    assert disabled_error["payload"]["context_api_enabled"] is False
-    assert disabled_error["payload"]["can_enable_context_api"] is True
-    assert enable_without_token["status"] == 403
-    assert settings_payload["schema"] == "codex-usage-tracker-context-settings-v1"
-    assert settings_payload["context_api_enabled"] is True
-    assert settings_payload["raw_context_persisted"] is False
-    assert usage_payload["context_api_enabled"] is True
-    assert context_payload["loaded_on_demand"] is True
-    assert context_payload["raw_context_persisted"] is False
-    assert context_payload["context_mode"] == "quick"
-    assert unlimited_context_payload["omitted"]["max_chars"] == 0
-    assert unlimited_context_payload["omitted"]["max_entries"] == 0
-    assert unlimited_context_payload["omitted"]["older_entries"] == 0
-
-
 def test_dashboard_query_limit_zero_loads_all_rows(tmp_path: Path) -> None:
     codex_home = _make_codex_home(tmp_path)
     db_path = tmp_path / "usage.sqlite3"
@@ -2000,463 +1145,3 @@ def test_dashboard_query_limit_zero_loads_all_rows(tmp_path: Path) -> None:
     assert len(query_dashboard_events(db_path=db_path, limit=2)) == 2
     assert len(query_dashboard_events(db_path=db_path, limit=0)) == 4
     assert query_dashboard_event_count(db_path=db_path) == 4
-
-
-def test_context_loads_raw_log_only_on_demand(tmp_path: Path) -> None:
-    codex_home = _make_codex_home(tmp_path)
-    db_path = tmp_path / "usage.sqlite3"
-    refresh_usage_index(codex_home=codex_home, db_path=db_path)
-    rows = query_session_usage(db_path=db_path, session_id=SESSION_ID)
-
-    context = load_call_context(rows[0]["record_id"], db_path=db_path)
-    context_text = json.dumps(context)
-
-    assert context["loaded_on_demand"] is True
-    assert context["raw_context_persisted"] is False
-    assert context["context_mode"] == "quick"
-    assert context["visible_char_count"] > 0
-    assert context["visible_token_estimate"] > 0
-    assert context["visible_token_estimator"] in {
-        "chars_per_4_fallback",
-        "tiktoken:o200k_base",
-        "tiktoken:cl100k_base",
-    }
-    serialized = context["serialized_evidence"]
-    assert serialized["available"] is True
-    assert serialized["scope"] == "selected_turn_raw_jsonl_fast_estimate"
-    assert serialized["upper_bound"] is True
-    assert serialized["raw_text_returned"] is False
-    assert serialized["deferred"] is True
-    assert serialized["deferred_buckets"] is True
-    assert serialized["reason"] == "full_serialized_analysis_not_requested"
-    assert serialized["raw_line_count"] >= len(context["entries"])
-    assert serialized["raw_json_char_count"] > context["visible_char_count"]
-    assert serialized["raw_json_token_estimate"] >= context["visible_token_estimate"]
-    assert serialized["token_estimator"] == "chars_per_4_fallback"
-    assert serialized["buckets"] == []
-    full_context = load_call_context(rows[0]["record_id"], db_path=db_path, mode="full")
-    full_serialized = full_context["serialized_evidence"]
-    assert full_context["context_mode"] == "full"
-    assert full_serialized["scope"] == "selected_turn_raw_jsonl"
-    assert full_serialized["deferred"] is False
-    assert full_serialized["deferred_buckets"] is False
-    assert full_serialized["token_estimator"] in {
-        "chars_per_4_fallback",
-        "tiktoken:o200k_base",
-        "tiktoken:cl100k_base",
-    }
-    serialized_bucket_keys = {bucket["key"] for bucket in full_serialized["buckets"]}
-    assert "encrypted_reasoning_state" in serialized_bucket_keys
-    assert "local_goal_metadata" in serialized_bucket_keys
-    assert "call_anchors" not in context
-    assert "thread_anchors" not in context
-    assert context["omitted"]["total_entries"] >= context["omitted"]["returned_entries"]
-    assert "SECRET RAW PROMPT" in context_text
-    assert "sk" + "-proj-" not in context_text
-    assert "AKIAIOSFODNN7EXAMPLE" not in context_text
-    assert "Authorization: Bearer abc.def" not in context_text
-    assert "xoxb-123456789012" not in context_text
-    assert "eyJhbGciOiJIUzI1Ni" not in context_text
-    assert "client_secret=super-secret-value" not in context_text
-    assert "BEGIN OPENSSH PRIVATE KEY" not in context_text
-    assert "[REDACTED_OPENAI_KEY]" in context_text
-    assert "[REDACTED_AWS_ACCESS_KEY]" in context_text
-    assert "[REDACTED_BEARER_TOKEN]" in context_text
-    assert "[REDACTED_SLACK_TOKEN]" in context_text
-    assert "[REDACTED_JWT]" in context_text
-    assert "[REDACTED_PRIVATE_KEY]" in context_text
-    assert any(entry["label"] == "message / user" for entry in context["entries"])
-    token_entry = next(entry for entry in context["entries"] if entry["label"] == "Token count")
-    assert token_entry["token_usage"]["last_token_usage"]["uncached_input_tokens"] >= 0
-    compaction_entries = [entry for entry in context["entries"] if entry["label"] == "Compaction detected"]
-    assert len(compaction_entries) == 2
-    compaction_entry = next(
-        entry for entry in compaction_entries if entry["compaction"]["replacement_entry_count"] == 2
-    )
-    event_compaction_entry = next(
-        entry for entry in compaction_entries if entry["compaction"]["replacement_entry_count"] == 1
-    )
-    assert compaction_entry["compaction"]["replacement_history_available"] is True
-    assert compaction_entry["compaction"]["replacement_entry_count"] == 2
-    assert compaction_entry["compaction"]["replacement_history_included"] is False
-    assert event_compaction_entry["compaction"]["replacement_history_available"] is True
-    assert event_compaction_entry["compaction"]["replacement_history_included"] is False
-    assert "replacement_history" not in compaction_entry["compaction"]
-    assert "COMPACTED REPLACEMENT SUMMARY" not in context_text
-    assert "EVENT MSG COMPACTION SUMMARY" not in context_text
-    assert "ENCRYPTED_STATE_SENTINEL_DO_NOT_RETURN" not in context_text
-    assert "LOCAL_GOAL_SENTINEL_DO_NOT_RETURN" not in context_text
-
-    compaction_context = load_call_context(
-        rows[0]["record_id"],
-        db_path=db_path,
-        include_compaction_history=True,
-    )
-    compaction_context_text = json.dumps(compaction_context)
-    compaction_with_history = next(
-        entry for entry in compaction_context["entries"] if entry["label"] == "Compaction detected"
-    )
-    assert compaction_context["include_compaction_history"] is True
-    assert compaction_with_history["compaction"]["replacement_history_included"] is True
-    assert "COMPACTED REPLACEMENT SUMMARY" in compaction_context_text
-    assert "EVENT MSG COMPACTION SUMMARY" in compaction_context_text
-    assert "sk" + "-proj-compactedsecret" not in compaction_context_text
-    assert "[REDACTED_OPENAI_KEY]" in compaction_context_text
-
-    limited_context = load_call_context(
-        rows[0]["record_id"],
-        db_path=db_path,
-        max_chars=40,
-        max_entries=1,
-    )
-    unlimited_context = load_call_context(
-        rows[0]["record_id"],
-        db_path=db_path,
-        max_chars=0,
-        max_entries=0,
-    )
-    assert limited_context["omitted"]["returned_entries"] == 1
-    assert limited_context["omitted"]["older_entries"] > 0
-    assert unlimited_context["omitted"]["max_chars"] == 0
-    assert unlimited_context["omitted"]["max_entries"] == 0
-    assert unlimited_context["omitted"]["older_entries"] == 0
-    assert unlimited_context["omitted"]["over_budget_chars"] == 0
-    assert len(unlimited_context["entries"]) > len(limited_context["entries"])
-
-
-def test_context_loading_uses_one_source_scan_for_evidence_and_serialized_estimate(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    codex_home = _make_codex_home(tmp_path)
-    db_path = tmp_path / "usage.sqlite3"
-    refresh_usage_index(codex_home=codex_home, db_path=db_path)
-    row = query_session_usage(db_path=db_path, session_id=SESSION_ID)[0]
-    source_file = Path(str(row["source_file"]))
-    open_count = 0
-    real_open = Path.open
-
-    def counting_open(path: Path, *args: object, **kwargs: object):
-        nonlocal open_count
-        if path == source_file:
-            open_count += 1
-        return real_open(path, *args, **kwargs)
-
-    monkeypatch.setattr(Path, "open", counting_open)
-
-    context = load_call_context(row["record_id"], db_path=db_path, diagnostics=True)
-
-    assert open_count == 1
-    assert any(entry["label"] == "message / user" for entry in context["entries"])
-    assert context["include_tool_output"] is False
-    assert context["context_mode"] == "quick"
-    assert context["serialized_evidence"]["available"] is True
-    assert context["serialized_evidence"]["deferred_buckets"] is True
-    assert "call_anchors" not in context
-    assert "thread_anchors" not in context
-    assert context["diagnostics"]["source_scan_ms"] >= 0
-    assert context["diagnostics"]["serialized_estimate_ms"] >= 0
-
-
-def test_context_carries_incoming_compaction_history_into_selected_turn(tmp_path: Path) -> None:
-    session_id = "019e37d5-f19f-7e4d-84cb-508941431111"
-    codex_home = tmp_path / ".codex"
-    log_path = (
-        codex_home
-        / "sessions"
-        / "2026"
-        / "06"
-        / "11"
-        / f"rollout-2026-06-11T22-20-00-{session_id}.jsonl"
-    )
-    _write_jsonl(
-        codex_home / "session_index.jsonl",
-        [
-            {
-                "id": session_id,
-                "thread_name": "Compaction boundary",
-                "updated_at": "2026-06-11T22:30:00Z",
-            }
-        ],
-    )
-    _write_jsonl(
-        log_path,
-        [
-            _entry("session_meta", {"id": session_id}),
-            _entry("turn_context", {"turn_id": "before-compact", "model": "gpt-5.5"}),
-            _token_event(100, 100),
-            _entry(
-                "compacted",
-                {
-                    "message": "",
-                    "replacement_history": [
-                        {
-                            "type": "message",
-                            "role": "assistant",
-                            "content": [
-                                {
-                                    "type": "output_text",
-                                    "text": "INCOMING REPLACEMENT HISTORY sk-proj-boundarysecret",
-                                }
-                            ],
-                        }
-                    ],
-                },
-            ),
-            _entry("turn_context", {"turn_id": "after-compact", "model": "gpt-5.5"}),
-            _entry("event_msg", {"type": "context_compacted"}),
-            _token_event(300, 200),
-        ],
-    )
-    db_path = tmp_path / "usage.sqlite3"
-    refresh_usage_index(codex_home=codex_home, db_path=db_path)
-    target = next(
-        row
-        for row in query_session_usage(db_path=db_path, session_id=session_id)
-        if row["turn_id"] == "after-compact"
-    )
-
-    context = load_call_context(
-        target["record_id"],
-        db_path=db_path,
-        include_compaction_history=True,
-    )
-    compaction_entries = [entry for entry in context["entries"] if entry["label"] == "Compaction detected"]
-
-    assert [entry["line_number"] for entry in compaction_entries] == [4, 6]
-    assert compaction_entries[0]["compaction"]["replacement_entry_count"] == 1
-    assert compaction_entries[0]["compaction"]["replacement_history_included"] is True
-    assert compaction_entries[1]["compaction"]["replacement_entry_count"] == 0
-    assert "Compaction marker found" in compaction_entries[1]["text"]
-    context_text = json.dumps(context)
-    assert "INCOMING REPLACEMENT HISTORY" in context_text
-    assert "sk-proj-boundarysecret" not in context_text
-    assert "[REDACTED_OPENAI_KEY]" in context_text
-
-
-def test_context_dedupes_adjacent_chat_message_echoes(tmp_path: Path) -> None:
-    session_id = "019e37d5-f19f-7e4d-84cb-508941432222"
-    codex_home = tmp_path / ".codex"
-    log_path = (
-        codex_home
-        / "sessions"
-        / "2026"
-        / "06"
-        / "11"
-        / f"rollout-2026-06-11T22-40-00-{session_id}.jsonl"
-    )
-    _write_jsonl(
-        codex_home / "session_index.jsonl",
-        [{"id": session_id, "thread_name": "Echo dedupe", "updated_at": "2026-06-11T22:45:00Z"}],
-    )
-    _write_jsonl(
-        log_path,
-        [
-            _entry("session_meta", {"id": session_id}),
-            _entry("turn_context", {"turn_id": "echo-turn", "model": "gpt-5.5"}),
-            _entry(
-                "response_item",
-                {
-                    "type": "message",
-                    "role": "user",
-                    "content": [{"type": "input_text", "text": "same user text"}],
-                },
-            ),
-            _entry("event_msg", {"type": "user_message", "message": "same user text"}),
-            _entry("event_msg", {"type": "agent_message", "message": "same assistant text"}),
-            _entry(
-                "response_item",
-                {
-                    "type": "message",
-                    "role": "assistant",
-                    "content": [{"type": "output_text", "text": "same assistant text"}],
-                },
-            ),
-            _token_event(300, 200),
-        ],
-    )
-    db_path = tmp_path / "usage.sqlite3"
-    refresh_usage_index(codex_home=codex_home, db_path=db_path)
-    target = query_session_usage(db_path=db_path, session_id=session_id)[0]
-
-    context = load_call_context(target["record_id"], db_path=db_path)
-    labels = [entry["label"] for entry in context["entries"]]
-    texts = [entry["text"] for entry in context["entries"]]
-
-    assert "message / user" in labels
-    assert "message / assistant" in labels
-    assert "user_message" not in labels
-    assert "agent_message" not in labels
-    assert texts.count("same user text") == 1
-    assert texts.count("same assistant text") == 1
-
-
-def test_mcp_wrappers_smoke(tmp_path: Path, monkeypatch) -> None:
-    from codex_usage_tracker import mcp_server
-
-    codex_home = _make_codex_home(tmp_path)
-    db_path = tmp_path / "usage.sqlite3"
-    dashboard_path = tmp_path / "dashboard.html"
-    pricing_path = _write_pricing(tmp_path / "pricing.json")
-    allowance_path = tmp_path / "allowance.json"
-    projects_path = tmp_path / "projects.json"
-    monkeypatch.setattr(mcp_server, "DEFAULT_CODEX_HOME", codex_home)
-    monkeypatch.setattr(mcp_server, "DEFAULT_DB_PATH", db_path)
-    monkeypatch.setattr(mcp_server, "DEFAULT_DASHBOARD_PATH", dashboard_path)
-    monkeypatch.setattr(mcp_server, "DEFAULT_PRICING_PATH", pricing_path)
-    monkeypatch.setattr(mcp_server, "DEFAULT_ALLOWANCE_PATH", allowance_path)
-    monkeypatch.setattr(mcp_server, "DEFAULT_PROJECTS_PATH", projects_path)
-    monkeypatch.setattr(mcp_server, "update_pricing_from_openai_docs", _fake_pricing_update)
-
-    refresh = mcp_server.refresh_usage_index()
-    summary = mcp_server.usage_summary(group_by="thread")
-    summary_json = mcp_server.usage_summary(group_by="model", response_format="json")
-    project_summary = mcp_server.usage_summary(group_by="project")
-    model_summary = mcp_server.usage_summary(preset="by-model")
-    expensive = mcp_server.most_expensive_usage_calls(limit=1)
-    expensive_json = mcp_server.most_expensive_usage_calls(limit=1, response_format="json")
-    query_json = mcp_server.usage_query(
-        model="gpt-5.5",
-        min_tokens=50,
-        limit=2,
-        privacy_mode="strict",
-    )
-    recommendations_json = mcp_server.usage_recommendations(
-        limit=2,
-        response_format="json",
-        privacy_mode="strict",
-    )
-    pricing_coverage = mcp_server.usage_pricing_coverage()
-    pricing_coverage_json = mcp_server.usage_pricing_coverage(response_format="json")
-    session = mcp_server.session_usage(session_id=SESSION_ID)
-    session_json = mcp_server.session_usage(session_id=SESSION_ID, response_format="json")
-    record_id = query_session_usage(db_path=db_path, session_id=SESSION_ID)[0]["record_id"]
-    context_disabled = mcp_server.usage_call_context(record_id=record_id)
-    context_disabled_json = json.loads(context_disabled)
-    monkeypatch.setenv("CODEX_USAGE_TRACKER_ALLOW_RAW_CONTEXT", "1")
-    context = mcp_server.usage_call_context(record_id=record_id)
-    context_json = json.loads(context)
-    dashboard = mcp_server.generate_usage_dashboard()
-    csv_export = mcp_server.export_usage_csv(str(tmp_path / "usage.csv"), privacy_mode="redacted")
-    pricing_init = mcp_server.init_usage_pricing_config(force=True)
-    pricing_update = mcp_server.update_usage_pricing_config()
-    allowance = mcp_server.init_usage_allowance_config()
-    doctor = mcp_server.usage_doctor()
-    doctor_json = mcp_server.usage_doctor(response_format="json")
-
-    for payload in (
-        refresh,
-        summary_json,
-        expensive_json,
-        query_json,
-        recommendations_json,
-        pricing_coverage_json,
-        session_json,
-        context_disabled_json,
-        context_json,
-        dashboard,
-        csv_export,
-        pricing_init,
-        pricing_update,
-        allowance,
-        doctor_json,
-    ):
-        _assert_contract(payload)
-
-    assert refresh["parsed_events"] == 4
-    assert refresh["skipped_events"] == 0
-    assert "Add Codex token tracking" in summary
-    assert summary_json["schema"] == "codex-usage-tracker-summary-v1"
-    assert summary_json["rows"][0]["group_key"] == "gpt-5.5"
-    assert "codex-usage-tracker" in project_summary
-    assert "estimated cost" in model_summary
-    assert "Most expensive Codex calls" in expensive
-    assert expensive_json["is_expensive"] is True
-    assert query_json["schema"] == "codex-usage-tracker-query-v1"
-    assert query_json["filters"]["model"] == "gpt-5.5"
-    assert query_json["row_count"] == 2
-    assert query_json["rows"][0]["pricing_model"] == "gpt-5.5"
-    assert query_json["rows"][0]["cwd"].startswith("[redacted cwd:")
-    assert query_json["rows"][0]["project_relative_cwd"] is None
-    assert recommendations_json["schema"] == "codex-usage-tracker-recommendations-v1"
-    assert recommendations_json["row_count"] >= 1
-    assert recommendations_json["rows"][0]["recommendation_score"] > 0
-    assert recommendations_json["threads"]
-    assert "Codex pricing coverage" in pricing_coverage
-    assert pricing_coverage_json["schema"] == "codex-usage-tracker-pricing-coverage-v1"
-    assert SESSION_ID in session
-    assert session_json["resolved_session_id"] == SESSION_ID
-    assert session_json["row_count"] == 2
-    assert "Raw context loading through MCP is disabled" in context_disabled
-    assert context_disabled_json["schema"] == "codex-usage-tracker-context-disabled-v1"
-    assert "SECRET RAW PROMPT" not in context_disabled
-    assert "SECRET RAW PROMPT" in context
-    assert context_json["schema"] == "codex-usage-tracker-context-v1"
-    assert "sk" + "-proj-" not in context
-    assert "[REDACTED_OPENAI_KEY]" in context
-    assert dashboard["dashboard_path"] == str(dashboard_path)
-    assert csv_export["privacy_mode"] == "redacted"
-    assert pricing_init["pricing_path"] == str(pricing_path)
-    assert pricing_update["model_count"] == 1
-    assert pricing_update["source_url"] == "https://example.test/pricing.md"
-    assert allowance["allowance_path"] == str(allowance_path)
-    assert allowance_path.exists()
-    assert "Codex Usage Tracker doctor" in doctor
-    assert doctor_json["schema"] == "codex-usage-tracker-doctor-v1"
-
-
-def test_pricing_annotation_and_doctor_pass(tmp_path: Path) -> None:
-    codex_home = _make_codex_home(tmp_path)
-    db_path = tmp_path / "usage.sqlite3"
-    dashboard_path = tmp_path / "dashboard.html"
-    pricing_path = _write_pricing(tmp_path / "pricing.json")
-    refresh_usage_index(codex_home=codex_home, db_path=db_path)
-    generate_dashboard(db_path=db_path, output_path=dashboard_path, pricing_path=pricing_path)
-
-    rows = query_most_expensive_calls(db_path=db_path, limit=1)
-    annotated = annotate_rows_with_efficiency(
-        rows, pricing=load_pricing_config(tmp_path / "missing-pricing.json")
-    )
-    assert annotated[0]["estimated_cost_usd"] is None
-    annotated = annotate_rows_with_efficiency(rows, pricing=load_pricing_config(pricing_path))
-    assert annotated[0]["estimated_cost_usd"] > 0
-
-    repo_root = tmp_path / "repo"
-    (repo_root / ".codex-plugin").mkdir(parents=True)
-    (repo_root / ".codex-plugin" / "plugin.json").write_text("{}", encoding="utf-8")
-    (repo_root / ".mcp.json").write_text(
-        json.dumps(
-            {
-                "mcpServers": {
-                    "codex-usage-tracker": {
-                        "command": sys.executable,
-                        "args": ["-m", "codex_usage_tracker.mcp_server"],
-                        "env": {
-                            "PYTHONPATH": str(Path(__file__).resolve().parents[1] / "src")
-                        },
-                    }
-                }
-            }
-        ),
-        encoding="utf-8",
-    )
-    plugin_link = tmp_path / "plugins" / "codex-usage-tracker"
-    plugin_link.parent.mkdir()
-    plugin_link.symlink_to(repo_root, target_is_directory=True)
-    marketplace_path = tmp_path / "marketplace.json"
-    marketplace_path.write_text(
-        json.dumps({"plugins": [{"name": "codex-usage-tracker"}]}),
-        encoding="utf-8",
-    )
-
-    report = run_doctor(
-        codex_home=codex_home,
-        db_path=db_path,
-        dashboard_path=dashboard_path,
-        pricing_path=pricing_path,
-        plugin_link=plugin_link,
-        marketplace_path=marketplace_path,
-        repo_root=repo_root,
-    )
-
-    assert report["status"] == "pass"
